requirements.txt
absl-py==1.0.0
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.0
astunparse==1.6.3
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
isort==5.10.1
keras==2.7.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
libclang==12.0.0
Markdown==3.3.6
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.8.2
pylint==2.12.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
